Se creo el calcularSueldoGlobal para el admin y se movio los casos de uso que quedaban en localModel a localService
